Catalogue and Frontend script updated

+++ roboshop/components/frontend.sh
@@ -1,10 +1,23 @@
 #!/bin/bash
-echo frontend setup............
+#set -e
 dnf install nginx -y
+echo frontend setup............
+if [ $? -ne 0 ]; then
+  echo "Nginx install Failed.."
+  exit
+fi
+
 systemctl enable nginx
 systemctl start nginx
 rm -rf /usr/share/nginx/html/*
+
 curl -o /tmp/frontend.zip https://roboshop-artifacts.s3.amazonaws.com/frontend.zip
+
+if [ $? -ne 0 ]; then
+  echo "Download Frontend Failed.."
+  exit
+fi
+
 cd /usr/share/nginx/html
 unzip /tmp/frontend.zip
